+++ src/App.js
@@ -1,67 +1,44 @@
-import React, { useState } from "react";
-import {
-  BrowserRouter as Router,
-  Route,
-  Routes,
-  Navigate,
-} from "react-router-dom";
+import React from "react";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+
+import Home from "./pages/Home";
+import NewFeatured from "./pages/NewFeatured";
+import Men from "./pages/Men";
+import Women from "./pages/Women";
+import Kids from "./pages/Kids";
+import Sale from "./pages/Sale";
+
+import Favourites from "./pages/Favourites";
+import Cart from "./pages/Cart";
+
 import Header from "./components/Header";
-import productsData from "./data/productsData";
-import ProductList from "./components/ProductList";
-import SalesList from "./components/SalesList";
 import Footer from "./components/Footer";
 import "./styles.css";
 
-const categories = ["Sportswear", "Shoes", "Accessories"];
+//redux
+import { Provider } from "react-redux";
+import store from "./redux/store";
 
 const App = () => {
-  const storedCategory = localStorage.getItem("currentCategory");
-  const [currentCategory, setCurrentCategory] = useState(null);
-
-  const handleCategoryChange = (category) => {
-    const newCategory = currentCategory === category ? null : category;
-    setCurrentCategory(newCategory);
-    localStorage.setItem("currentCategory", newCategory);
-  };
-
   return (
-    <Router>
-      <div className="app">
-        <Routes>
-          <Route path="/" element={<Navigate to="/home" />} />
-        </Routes>
-        <Header
-          categories={categories}
-          currentCategory={currentCategory}
-          onCategoryChange={handleCategoryChange}
-        />
-        <Routes>
-          <Route
-            path="/home"
-            element={<ProductList category={null} loadAllProducts={true} />}
-          ></Route>
-          <Route
-            path="/sale"
-            element={
-              <SalesList category={null} loadDiscountedProducts={true} />
-            }
-          ></Route>
-          <Route
-            path={`/home/:category?`}
-            element={
-              <ProductList products={productsData} category={storedCategory} />
-            }
-          />
-          <Route
-            path={`/sale/:category?`}
-            element={
-              <SalesList products={productsData} category={storedCategory} />
-            }
-          />
-        </Routes>
-        <Footer />
-      </div>
-    </Router>
+    <Provider store={store}>
+      <Router>
+        <div className="app">
+          <Header />
+          <Routes>
+            <Route path="/" element={<Home />} />
+            <Route path="/NewFeatured" element={<NewFeatured />} />
+            <Route path="/Men" element={<Men />} />
+            <Route path="/Women" element={<Women />} />
+            <Route path="/Kids" element={<Kids />} />
+            <Route path="/Sale" element={<Sale />} />
+            <Route path="/Favourites" element={<Favourites />} />
+            <Route path="/Cart" element={<Cart />} />
+          </Routes>
+          <Footer />
+        </div>
+      </Router>
+    </Provider>
   );
 };
