+++ src/components/Header.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+import NavBar from "./NavBar";
+import "./styles/Header.css";
+
+function Header() {
+  return (
+    <div>
+      <NavBar />
+    </div>
+  );
+}
+
+export default Header;
