--- src/components/styles/Header.css
@@ -1,91 +1,0 @@
-/* src/styles.css */
-.header {
-  background-color: white;
-  position: sticky;
-  top: 0;
-  z-index: 1000;
-  height: 15vh;
-}
-
-.nav-menu {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 1% 0;
-}
-
-.logo {
-  width: 100px;
-  height: 30px;
-}
-
-.main-link {
-  flex: 2; /* 60% for logo + links (3/5) */
-  text-align: center;
-  text-decoration: none;
-  color: #111111;
-  font-size: 1.3rem;
-}
-
-.new {
-  margin-left: 8%;
-}
-
-.nav-icon-container {
-  flex: 3; /* 40% for icons (2/5) */
-  display: flex;
-  justify-content: center; /* Align icons to the right */
-  gap: 5%;
-}
-
-.nav-icon {
-  font-size: 1.3rem;
-  border: none;
-  background: none;
-  cursor: pointer;
-}
-
-.search-input {
-  position: relative;
-  width: 100px;
-  padding-right: 10%;
-  margin-right: 5%;
-}
-
-.search-input input {
-  background-color: #f5f5f5;
-  border: 1px transparent;
-  border-radius: 20px;
-  font-size: 1rem;
-  padding: 5% 0 5% 40px;
-  width: 100%;
-}
-
-.search-icon {
-  position: absolute;
-  top: 60%;
-  left: 10px;
-  transform: translateY(-50%);
-  border: none;
-  background-color: transparent;
-  font-size: 1.3rem;
-}
-
-.categories-menu-container {
-  background-color: #f5f5f5;
-  padding: 0.5% 0 0.5% 0;
-}
-
-.categories-menu {
-  width: 40%;
-  margin: 0 30%;
-}
-
-@media only screen and (max-width: 632px) {
-  .nav-menu {
-    display: none;
-  }
-  .categories-menu-container {
-    display: none;
-  }
-}
