+++ src/components/styles/NavBar.css
@@ -1,0 +1,122 @@
+/* NavBar.css */
+.acc-link-container {
+  position: relative;
+  display: flex;
+  width: 100%;
+  z-index: 2;
+  height: 30px;
+  background-color: #F5F5F5;
+}
+
+.not-fixed {
+  position: relative;
+  width: 100%;
+  transition: transform 0.3s ease-in-out;
+  background-color: white;
+}
+
+.fixed {
+  position: fixed;
+  z-index: 1;
+  top: 0;
+  width: 100%;
+  transition: transform 0.3s ease-in-out;
+  background-color: white;
+}
+
+.visible {
+  transform: translateY(0%);
+}
+
+.hidden {
+  transform: translateY(-120%);
+}
+
+.nav-wrapper {
+  padding: 0.5% 0%;
+  background-color: #fff;
+}
+
+.nav-container {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.logo {
+  width: 100px;
+  height: 30px;
+  fill: #111111;
+  transition: fill 0.2s ease-in-out;
+  margin-left: 25%;
+}
+
+.logo:hover {
+  fill: #7e7e7e;
+}
+
+.nav-link-container {
+  margin-left: 15%;
+  display: flex;
+}
+
+.nav-link {
+  font-weight: 600;
+  text-decoration: none;
+  color: #000000;
+  border-bottom: 2px solid transparent;
+  padding: 10px;
+}
+
+.nav-link:hover {
+  border-bottom: 2px solid black;
+}
+
+.nav-icon-container {
+  display: flex;
+  gap: 10px;
+  margin-right: 2%;
+}
+
+.search-input {
+  position: relative;
+  width: 200px;
+  padding-right: 10%;
+  margin-right: 5%;
+}
+
+.search-input input {
+  background-color: #f5f5f5;
+  border: 1px transparent;
+  border-radius: 20px;
+  font-size: 1rem;
+  padding: 7% 0 5% 40px;
+  margin: 0% 0 0 0;
+  width: 100%;
+}
+
+.search-input input:hover {
+  background: #00000020;
+}
+
+.search-icon {
+  position: absolute;
+  top: 50%;
+  left: 0px;
+  transform: translateY(-50%);
+  border: none;
+  background-color: transparent;
+  font-size: 1.5rem;
+  border-radius: 50%;
+}
+
+.nav-icon {
+  font-size: 1.5rem;
+  background: none;
+  cursor: pointer;
+  color: #000000;
+  border-radius: 50%;
+  padding: 1.5% 3.75%;
+}
+
+/* Add additional styles as needed */
