+++ src/components/styles/VideoPlayer.css
@@ -1,0 +1,4 @@
+.video-player {
+    width: 100%;
+    margin-top: 200px;
+}
