+++ src/pages/Cart.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Cart() {
+  return (
+    <div>Cart</div>
+  )
+}
+
+export default Cart
