+++ src/pages/Favourites.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Favourites() {
+  return (
+    <div>Favourites</div>
+  )
+}
+
+export default Favourites
