+++ src/pages/Football.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Football() {
+  return (
+    <div>Football</div>
+  )
+}
+
+export default Football
