+++ src/pages/Home.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+import VideoPlayer from "../components/VideoPlayer";
+
+function Home() {
+  return (
+    <div>
+      <VideoPlayer />
+    </div>
+  );
+}
+
+export default Home;
