+++ src/pages/Kids.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Kids() {
+  return (
+    <div>Kids</div>
+  )
+}
+
+export default Kids
