+++ src/pages/Men.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Men() {
+  return (
+    <div>Men</div>
+  )
+}
+
+export default Men
