+++ src/pages/NewFeatured.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function NewFeatured() {
+  return (
+    <div>NewFeatured</div>
+  )
+}
+
+export default NewFeatured
