+++ src/pages/Sale.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Sale() {
+  return (
+    <div>Sale</div>
+  )
+}
+
+export default Sale
