+++ src/pages/Women.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Women() {
+  return (
+    <div>Women</div>
+  )
+}
+
+export default Women
