+++ src/redux/index.js
@@ -1,0 +1,1 @@
+export { SET_NAVBAR_VISIBILITY } from "./nav/navActions";
