+++ src/redux/rootReducer.js
@@ -1,0 +1,6 @@
+import { combineReducers } from "redux";
+
+const rootReducer = combineReducers({
+})
+
+export default rootReducer;
